bosung-xay_dung_chuc_nang_CRUD_du_lieu

+++ DEMO_MVC/Controllers/PersonController.cs
@@ -14,18 +14,37 @@ namespace DEMO_MVC.Controllers
             _context = context;
         }
 
-        // Action Index: trả về View list Person trong CSDL
+        // GET: Person
         public async Task<IActionResult> Index()
         {
-            var model = await _context.Person.ToListAsync();
-            return View(model);
+            return View(await _context.Person.ToListAsync());
         }
 
+        // GET: Person/Details/5
+        public async Task<IActionResult> Details(string id)
+        {
+            if (id == null || _context.Person == null)
+            {
+                return NotFound();
+            }
+
+            var person = await _context.Person
+                .FirstOrDefaultAsync(m => m.PersonId == id);
+            if (person == null)
+            {
+                return NotFound();
+            }
+
+            return View(person);
+        }
+
+        // GET: Person/Create
         public IActionResult Create()
         {
             return View();
         }
 
+        // POST: Person/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("PersonId,FullName,Address")] Person person)
@@ -39,19 +58,31 @@ namespace DEMO_MVC.Controllers
             return View(person);
         }
 
+        // GET: Person/Edit/5
         public async Task<IActionResult> Edit(string id)
         {
-            if (id == null) return NotFound();
+            if (id == null || _context.Person == null)
+            {
+                return NotFound();
+            }
+
             var person = await _context.Person.FindAsync(id);
-            if (person == null) return NotFound();
+            if (person == null)
+            {
+                return NotFound();
+            }
             return View(person);
         }
 
+        // POST: Person/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(string id, [Bind("PersonId,FullName,Address")] Person person)
         {
-            if (id != person.PersonId) return NotFound();
+            if (id != person.PersonId)
+            {
+                return NotFound();
+            }
 
             if (ModelState.IsValid)
             {
@@ -62,38 +93,61 @@ namespace DEMO_MVC.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!PersonExists(person.PersonId)) return NotFound();
-                    else throw;
+                    if (!PersonExists(person.PersonId))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
                 }
                 return RedirectToAction(nameof(Index));
             }
             return View(person);
         }
 
+        // GET: Person/Delete/5
         public async Task<IActionResult> Delete(string id)
         {
-            if (id == null) return NotFound();
-            var person = await _context.Person.FirstOrDefaultAsync(m => m.PersonId == id);
-            if (person == null) return NotFound();
+            if (id == null || _context.Person == null)
+            {
+                return NotFound();
+            }
+
+            var person = await _context.Person
+                .FirstOrDefaultAsync(m => m.PersonId == id);
+            if (person == null)
+            {
+                return NotFound();
+            }
+
             return View(person);
         }
 
+        // POST: Person/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            if (_context.Person == null)
+            {
+                return Problem("Entity set 'ApplicationDbContext.Person' is null.");
+            }
             var person = await _context.Person.FindAsync(id);
             if (person != null)
             {
                 _context.Person.Remove(person);
-                await _context.SaveChangesAsync();
             }
+
+            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
+        // Kiểm tra Person có tồn tại trong CSDL không
         private bool PersonExists(string id)
         {
-            return _context.Person.Any(e => e.PersonId == id);
+            return (_context.Person?.Any(e => e.PersonId == id)).GetValueOrDefault();
         }
     }
 }
